backend/pyproject.toml
[project]
name = "prompts-volume"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "fastapi>=0.121.0",
    "hdbscan>=0.8.33",
    "httpx>=0.28.1",
    "openai>=2.7.1",
    "pgvector>=0.4.1",
    "pydantic-settings>=2.11.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.3.0",
    "sentence-transformers>=3.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "testcontainers[postgres]>=4.13.3",
    "torch>=2.0.0",
    "uvicorn>=0.38.0",
    "pyjwt>=2.8.0",
    "python-multipart>=0.0.9",
    "email-validator>=2.3.0",
    "bcrypt>=5.0.0",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cpu" }
torchvision = { index = "pytorch-cpu" }
torchaudio = { index = "pytorch-cpu" }
